Replaced super inefficient usage of apc_cache_info('user') with APCIterator.
If the APC user cache is very large, apc_cache_info('user') can lock up the server. See https://bugs.php.net/bug.php?id=59937

Note that APCIterator was first available in APC 3.1.1. If you want to include support for versions of APC before 3.1.1, you could easily check for the availability of the APCIterator class and fall back to the old code, but I don't recommend that due to the possibility of it locking up the server.